Some Screen navigations done Added some AppRoutes objects Added RecipientTextField.kt which is used in amount screen and favourite edit Favourite Bottom Sheet added to AmountScreen made some changes 3ndy fel more more and favourite Screens
