+++ app/src/main/java/com/belia/speedotransfer/ui/main_screens/more/MoreScreen.kt
@@ -3,20 +3,14 @@ package com.belia.speedotransfer.ui.main_screens.more
 import android.content.Intent
 import android.net.Uri
 import androidx.annotation.DrawableRes
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.material3.CardDefaults
-import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
@@ -40,15 +34,14 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
 import com.belia.speedotransfer.R
+import com.belia.speedotransfer.navigation.AppRoutes
 import com.belia.speedotransfer.navigation.AppRoutes.FAVOURITES
 import com.belia.speedotransfer.ui.common_ui.SpeedoNavigationBar
 import com.belia.speedotransfer.ui.common_ui.TopBar
+import com.belia.speedotransfer.ui.main_screens.more.help.HelpBottomSheet
 import com.belia.speedotransfer.ui.theme.GrayG200
 import com.belia.speedotransfer.ui.theme.GrayG40
-import com.belia.speedotransfer.ui.theme.GrayG900
-import com.belia.speedotransfer.ui.theme.RedP300
 import com.belia.speedotransfer.ui.theme.bodyMedium16
-import com.belia.speedotransfer.ui.theme.bodyRegular16
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -58,6 +51,7 @@ fun MoreScreen(
 ) {
     var isBottomSheetVisible by remember { mutableStateOf(false) }
     var state = rememberModalBottomSheetState()
+    var context = LocalContext.current
 
     Scaffold(
         topBar = {
@@ -77,7 +71,10 @@ fun MoreScreen(
                     )
                 )
         ) {
-            MoreItem(icon = R.drawable.website, title = "Transfer From Website", hasSahm = true)
+            MoreItem(icon = R.drawable.website, title = "Transfer From Website", hasSahm = true){
+                val intent = Intent(Intent.ACTION_VIEW, Uri.parse("https://www.google.com"))
+                context.startActivity(intent)
+            }
             HorizontalDivider(modifier = modifier.padding(horizontal = 16.dp), color = GrayG40)
             MoreItem(icon = R.drawable.favorite, title = "Favorites", hasSahm = true) {
                 navController.navigate(FAVOURITES)
@@ -89,7 +86,10 @@ fun MoreScreen(
                 isBottomSheetVisible = true
             }
             HorizontalDivider(modifier = modifier.padding(horizontal = 16.dp), color = GrayG40)
-            MoreItem(icon = R.drawable.logout, title = "Logout")
+            MoreItem(icon = R.drawable.logout, title = "Logout"){
+                /*TODO (End API session)*/
+                navController.popBackStack(AppRoutes.LOGIN, inclusive = false)
+            }
         }
     }
 
@@ -104,87 +104,6 @@ fun MoreScreen(
         )
     }
 }
-
-@Composable
-fun HelpBottomSheet(modifier: Modifier = Modifier) {
-    val context = LocalContext.current
-    Spacer(modifier = modifier.height(56.dp))
-    Row(
-        modifier = modifier
-            .fillMaxWidth()
-            .padding(horizontal = 51.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.SpaceBetween
-    ) {
-        ElevatedCard(
-            elevation = CardDefaults.cardElevation(
-                defaultElevation = 3.dp
-            ),
-            onClick = {
-                val emailIntent = Intent(Intent.ACTION_SENDTO).apply {
-                    data = Uri.parse("mailto:support@example.com")
-                    putExtra(Intent.EXTRA_EMAIL, arrayOf("support@example.com"))
-                    putExtra(Intent.EXTRA_SUBJECT, "Support Request")
-                    putExtra(Intent.EXTRA_TEXT, "I need help with...")
-                }
-                context.startActivity(Intent.createChooser(emailIntent, "Send Email"))
-            },
-            colors = CardDefaults.cardColors(
-                containerColor = Color.White,
-            ),
-            modifier = modifier
-                .size(120.dp, 140.dp)
-        ) {
-            Column(
-                modifier = modifier.fillMaxSize(),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.email_square),
-                    contentDescription = "",
-                )
-                Spacer(modifier = modifier.height(16.dp))
-                Text(
-                    text = "Send Email",
-                    style = bodyMedium16,
-                    color = GrayG900,
-                    modifier = modifier.padding(bottom = 16.dp)
-                )
-            }
-        }
-        ElevatedCard(
-            elevation = CardDefaults.cardElevation(defaultElevation = 3.dp,),
-            onClick = {
-                val callIntent = Intent(Intent.ACTION_DIAL).apply {
-                    data = Uri.parse("tel:000000")
-                }
-                context.startActivity(callIntent)
-            },
-            colors = CardDefaults.cardColors(
-                containerColor = Color.White,
-            ),
-            modifier = modifier
-                .size(120.dp, 140.dp)
-        ) {
-            Column(
-                modifier = modifier.fillMaxSize(),
-                verticalArrangement = Arrangement.Center,
-                horizontalAlignment = Alignment.CenterHorizontally
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.callus_square),
-                    contentDescription = "",
-                )
-                Spacer(modifier = modifier.height(16.dp))
-                Text(text = "Call Us", style = bodyMedium16, color = GrayG900)
-                Text(text = "000000", style = bodyRegular16, color = RedP300)
-            }
-        }
-    }
-    Spacer(modifier = modifier.height(56.dp))
-}
-
 
 @Composable
 fun MoreItem(
