+++ app/src/main/java/com/belia/speedotransfer/ui/splash_screen/SplashScreen.kt
@@ -17,12 +17,15 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.sp
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
+import com.belia.speedotransfer.navigation.AppRoutes
 import com.belia.speedotransfer.ui.theme.InterFontFamily
 import com.belia.speedotransfer.ui.theme.RedP500
 import kotlinx.coroutines.delay
 
 @Composable
-fun SplashScreen() {
+fun SplashScreen(navController: NavController) {
     val scale = remember { Animatable(0.3f) }
 
     LaunchedEffect(key1 = true) {
@@ -34,7 +37,7 @@ fun SplashScreen() {
             )
         )
         delay(2000L)
-        // TODO: navigate to other screens
+        navController.navigate(AppRoutes.LOGIN)
     }
 
     Box(
@@ -58,5 +61,5 @@ fun SplashScreen() {
 @Preview(showBackground = true)
 @Composable
 private fun SplashScreenPreview() {
-    SplashScreen()
+    SplashScreen(rememberNavController())
 }
